worked on navbar, menu, header/scroll, combining files for less clutter and organization
