+++ client/src/App.js
@@ -1,20 +1,21 @@
 import './App.css';
 import "./assets/css/styles.css";
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+// import { Link } from 'react-router-dom';
 import { ApolloClient, InMemoryCache, ApolloProvider, createHttpLink } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
 import { StoreProvider } from './utils/GlobalState';
 
-import Header from './components/Header';
+import Nav from './components/Nav/Nav.js';
 import Menu from './components/Menu';
-import Item from './components/Item';
+// import Item from './components/Item';
 import Home from './pages/Home';
 import Login from './pages/Login';
 import Signup from './pages/Signup';
 import Footer from './components/Footer';
 import CafeTables from './pages/CafeTables';
 
-import Nav from './components/Nav';
+// import Nav from './components/Nav';
 
 const httpLink = createHttpLink({
   uri: '/graphql',
@@ -42,7 +43,7 @@ function App() {
       <Router>
     <div>
       <StoreProvider>
-        <Header />
+        <Nav name="Cafe Du Lait"/>
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/login" element={<Login />} />
@@ -50,11 +51,7 @@ function App() {
           <Route path="/menu" element={<Menu />} />
           <Route exact path="/cafe-tables" element={<CafeTables />} />
         </Routes>
-      {/* <Header name="Cafe Du Lait"/>
-      <Menu name="Cold Drinks" menu={Item}/>
-      <Menu name="Hot Drinks" menu={Item}/> */}
       </StoreProvider>
-   
     </div>
     <Footer />
     </Router>
