+++ client/src/assets/css/styles.css
@@ -24,12 +24,49 @@ h3 {
     color:rgb(7, 45, 34)
 }
 
-/* @font-face {
-    font-family: 'SoDo Sans';
-    src: local('SoDo Sans Thin'), local('SoDo-Sans-Thin'),
-        url('SoDoSans-Thin.woff2') format('woff2'),
-        url('SoDoSans-Thin.woff') format('woff'),
-        url('SoDoSans-Thin.ttf') format('truetype');
-    font-weight: 100;
-    font-style: normal;
+  /* .slider-wrapper {
+    position: relative;
+    top: 70px;
   } */
+
+  .flex-row {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+
+  .category-menu {
+    text-align: center;
+    margin-top: 50px;
+  }
+  
+  .category-buttons {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    
+  }
+  
+  .category-buttons button {
+    background-color: rgba(7, 45, 34, 0.719) !important;
+  border: none !important;
+  color: white !important;
+    margin: 10px;
+    padding: 10px 20px;
+    font-size: 18px;
+  }
+  
+  @media (max-width: 768px) {
+    .category-buttons {
+      flex-direction: column;
+      align-items: center;
+    }
+  }
+  
+  .category-buttons button:hover {
+    background-color: rgba(7, 45, 34, 1) !important;
+    color: white !important;
+  }
+  
+  
+
