+++ client/src/components/ItemList/index.js
@@ -58,7 +58,6 @@ function ProductList() {
               name={product.name}
               description={product.description}
               price={product.price}
-              quantity={product.quantity}
             />
           ))}
         </div>
